Move PushDecoderStreamState into push_decoder.rs
Co-locate the per-file stream driver with the builder configuration it
consumes. Add `into_stream` so the opener doesn't need to name the
unfold/fuse type; it just hands off the state and gets back a
`BoxStream`.

After this commit, push_decoder.rs owns the full push-decoder
lifecycle (builder setup + stream driving), and build_stream in
opener.rs reads as orchestration: prepare access plans, build
decoders, hand off to the stream, optionally wrap in EarlyStoppingStream.

No behavior change.
